// packages/runtime/runtime-extensions/src/request-policy/types.ts
export const BFF_ENVELOPE_HEADER = 'x-modernjs-bff-envelope';
export const BFF_OPERATION_CONTEXT_HEADER = 'x-operation-id';
export const BFF_OPERATION_CONTEXT_DETAIL_HEADER =
  'x-modernjs-bff-operation-context';
// ...
export type TransportTarget = 'server' | 'browser';